requirements.txt
semantic-kernel>=0.9.0
PyGithub>=2.0.0
